requirements.txt
Flask
python-dotenv
discord.py==2.3.2

aiofiles==23.2.1
gspread==6.0.2
gspread-asyncio==2.0.0
oauth2client==4.1.3

google-auth==2.25.2
google-api-python-client==2.128.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
